# ms-gateway/docs/swagger.yaml
definitions:
...
  helper.Message:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: kiet123pascal@gmail.com
    name: Kiet Asmara
    url: http://www.swagger.io/support
  description: E-commerce API with product management and shipment features.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Kiwish
  version: "1.0"
paths:
...
  /order/notification:
    get:
      consumes:
      - application/json
      description: Get All Notification
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Message'
      summary: Get Notification
      tags:
      - Order
  /order/notification/mark:
    put:
      consumes:
      - application/json
      description: Mark All notification as read
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Message'
      summary: Mark All Notification
      tags:
      - Order
  /order/notification/mark/:id:
    put:
      consumes:
      - application/json
      description: Mark Notification By id
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order ID
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Message'
      summary: Mark Notification By Id
      tags:
      - Order
...
swagger: "2.0"